+++ src/compas_timber/connections/l_butt.py
@@ -1,6 +1,7 @@
 from compas.geometry import Frame
 from compas_timber.parts import CutFeature
 
+from .joint import BeamJoinningError
 from .joint import Joint
 from .joint import beam_side_incidence
 from .solver import JointTopology
@@ -77,15 +78,13 @@ class LButtJoint(Joint):
     def joint_type(self):
         return "L-Butt"
 
-    @property
-    def cutting_plane_main(self):
+    def get_main_cutting_plane(self):
         angles_faces = beam_side_incidence(self.main_beam, self.cross_beam)
         cfr = min(angles_faces, key=lambda x: x[0])[1]
         cfr = Frame(cfr.point, cfr.xaxis, cfr.yaxis * -1.0)  # flip normal
         return cfr
 
-    @property
-    def cutting_plane_cross(self):
+    def get_cross_cutting_plane(self):
         angles_faces = beam_side_incidence(self.cross_beam, self.main_beam)
         cfr = max(angles_faces, key=lambda x: x[0])[1]
         return cfr
@@ -101,16 +100,24 @@ class LButtJoint(Joint):
         This method is automatically called when joint is created by the call to `Joint.create()`.
 
         """
+        assert self.main_beam and self.cross_beam  # should never happen
+
         if self.features:
             self.main_beam.remove_features(self.features)
 
-        start_main, end_main = self.main_beam.extension_to_plane(self.cutting_plane_main)
-        start_cross, end_cross = self.cross_beam.extension_to_plane(self.cutting_plane_cross)
+        try:
+            main_cutting_plane = self.get_main_cutting_plane()
+            cross_cutting_plane = self.get_cross_cutting_plane()
+        except Exception as ex:
+            raise BeamJoinningError(beams=self.beams, joint=self, debug_info=str(ex))
+
+        start_main, end_main = self.main_beam.extension_to_plane(main_cutting_plane)
+        start_cross, end_cross = self.cross_beam.extension_to_plane(cross_cutting_plane)
         self.main_beam.add_blank_extension(start_main, end_main, self.key)
         # self.cross_beam.add_blank_extension(start_cross, end_cross, self.key)
         self.cross_beam.add_blank_extension(1, 1, self.key)
 
-        f_main = CutFeature(self.cutting_plane_main)
+        f_main = CutFeature(main_cutting_plane)
         self.main_beam.add_features(f_main)
         self.features.append(f_main)