+++ src/compas_timber/consumers/__init__.py
@@ -1,7 +1,9 @@
 from .geometry import BeamGeometry
 from .geometry import BrepGeometryConsumer
+from .geometry import FeatureApplicationError
 
 __all__ = [
     "BrepGeometryConsumer",
     "BeamGeometry",
+    "FeatureApplicationError",
 ]